requirements.txt
python-dotenv==0.17.0
requests==2.26.0
geocoder==1.38.1